+++ loopback/backened_torneo/src/services/autentificacion.service.ts
@@ -3,8 +3,6 @@ import { repository } from '@loopback/repository';
 import { DirectorTecnico } from '../models';
 import { DirectorTecnicoRepository } from '../repositories';
 import { Llaves } from '../config/llaves';
-const generador = require("password-generator");
-const cryptoJS = require("crypto-js");
 const jwt = require("jsonwebtoken");
 
 
@@ -12,20 +10,12 @@ const jwt = require("jsonwebtoken");
 @injectable({scope: BindingScope.TRANSIENT})
 export class AutentificacionService {
   constructor(
-    @repository(DirectorTecnicoRepository) 
+    @repository(DirectorTecnicoRepository)
     public directorTecnicoRepository: DirectorTecnicoRepository
   ){}
-  generarClave(){
-    let clave = generador(8, false);
-    return clave;
-  }
-  
-  cifrarClave(clave: string){
-    let claveCifrada = cryptoJS.MD5(clave).toString();
-    return claveCifrada;
-  }
-  
-  
+
+
+
   IdentificarDirectorTecnico(usuario: string, Clave: string){
     try{
       let d = this.directorTecnicoRepository.findOne({where:{Correo: usuario, Clave: Clave}});
@@ -33,13 +23,13 @@ export class AutentificacionService {
                return d;
            }
            return false;
-  
+
        }catch{
-  
+
            return false;
          }
-      }  
-  
+      }
+
   GenerarTokenJWT(directorTecnico: DirectorTecnico){
     let token = jwt.sign({
       data:{
@@ -50,17 +40,17 @@ export class AutentificacionService {
       // TipoDocumento : directorTecnico.TipoDocumento,
       // Documento: directorTecnico.Documento
     }
-  
+
   },
       Llaves.claveJWT);
       return token;
   }
-  
+
   ValidarTokenJWT(token: string){
     try{
       let datos = jwt.verify(token, Llaves.claveJWT);
       return datos;
-  
+
     }catch{
       return false;
    }
